GestorEstoqueAI/requirements.txt
pandas
scikit-learn
streamlit
matplotlib
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
streamlit